+++ src/utils/utils.go
@@ -33,6 +33,7 @@ var NeedsRestart = false
 var UpdateAvailable = map[string]bool{}
 
 var RestartHTTPServer func()
+var ReBootstrapContainer func(string) error
 
 var LetsEncryptErrors = []string{}